implementing UpdateLayerStatistics

+++ test/check_shp_load_3d.c
@@ -54,6 +54,10 @@ int main (int argc, char *argv[])
     sqlite3 *handle;
     char *err_msg = NULL;
     int row_count;
+    const char *sql;
+    char **results;
+    int rows;
+    int columns;
 
     spatialite_init (0);
     ret = sqlite3_open_v2 (":memory:", &handle, SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE, NULL);
@@ -79,6 +83,13 @@ int main (int argc, char *argv[])
 	return -3;
     }
 
+    ret = update_layer_statistics (handle, "polygons", "geom");
+    if (!ret) {
+        fprintf (stderr, "update_layer_statistics() error\n", err_msg);
+	sqlite3_close(handle);
+	return -8;
+    }
+
     ret = load_shapefile (handle, "shp/merano-3d/roads", "roads", "CP1252", 25832, 
 			  "geom", 0, 0, 1, 0, &row_count, err_msg);
     if (!ret) {
@@ -98,6 +109,41 @@ int main (int argc, char *argv[])
     elementary_geometries (handle, "points", "geom", "elem_point", "pk_elem", "mul_id");
     elementary_geometries (handle, "roads", "geom", "elem_linestring", "pk_elem", "mul_id");
     elementary_geometries (handle, "polygons", "geom", "elem_poly", "pk_elem", "mul_id");
+
+    ret = update_layer_statistics (handle, NULL, NULL);
+    if (!ret) {
+        fprintf (stderr, "update_layer_statistics() error\n", err_msg);
+	sqlite3_close(handle);
+	return -8;
+    }
+
+    sql = "SELECT row_count, extent_min_x, extent_max_y "
+          "FROM layer_statistics "
+          "WHERE raster_layer = 0 AND table_name LIKE 'roads' "
+          "AND geometry_column LIKE 'geom'";
+    ret = sqlite3_get_table (handle, sql, &results, &rows, &columns, &err_msg);
+    if (ret != SQLITE_OK) {
+        fprintf (stderr, "Error: %s\n", err_msg);
+        sqlite3_free (err_msg);
+        return -9;
+    }
+    if (rows != 1) {
+        fprintf (stderr, "Unexpected num of rows for LAYER_STATISTICS.\n");
+        return  -10;
+    }
+    if (atoi(results[columns+0]) != 18) {
+        fprintf (stderr, "Unexpected error: LAYER_STATISTICS bad result row_count: %s\n", results[columns+0]);
+        return -11;
+    }
+    if (atof(results[columns+1]) != 666057.648017325) {
+        fprintf (stderr, "Unexpected error: LAYER_STATISTICS bad result extent_min_x: %s\n", results[columns+1]);
+        return -12;
+    }
+    if (atof(results[columns+2]) != 5170671.31627796) {
+        fprintf (stderr, "Unexpected error: LAYER_STATISTICS bad result extent_max_y: %s\n", results[columns+2]);
+        return -13;
+    }
+    sqlite3_free_table (results);
 
     ret = sqlite3_exec (handle, "DROP TABLE polygons", NULL, NULL, &err_msg);
     if (ret != SQLITE_OK) {
